requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
gspread
imap-tools
pandas
openpyxl
pdfplumber
python-dotenv
xlrd